extract question types and made them into components
