+++ src/components/ApplicationQuestions/SelectAll.jsx
@@ -1,0 +1,73 @@
+import { ErrorMessage } from '../Typography'
+import { applyCustomSort } from '../../utility/utilities'
+import { TextInput, Select } from '../Input'
+
+const SelectAll = ({ refs, errors, formInputs, onChange, question }) => {
+  const toCamelCase = str => {
+    return str
+      .toLowerCase()
+      .split(' ')
+      .map((word, index) => (index === 0 ? word : word.charAt(0).toUpperCase() + word.slice(1)))
+      .join('')
+  }
+
+  const toOtherCamelCase = str => {
+    const capitalizedStr = str.charAt(0).toUpperCase() + str.slice(1)
+    return `other${capitalizedStr}`
+  }
+
+  const transformSelectAllOptions = (options, includeOther) => {
+    const transformedOptions = options.reduce((acc, option) => {
+      acc[toCamelCase(option)] = option.toString()
+      return acc
+    }, {})
+
+    if (includeOther) {
+      transformedOptions.other = 'Other (Please specify)'
+    }
+
+    return transformedOptions
+  }
+
+  const selectAllOptions = transformSelectAllOptions(question.options, question.other)
+
+  return (
+    <>
+      {errors?.[question.formInput] && <ErrorMessage>{errors?.[question.formInput]}</ErrorMessage>}
+      {formInputs &&
+        applyCustomSort(
+          Object.entries(formInputs?.[question.formInput]),
+          Object.keys(selectAllOptions)
+        ).map(([key, val]) => (
+          <Select
+            key={key}
+            type="checkbox"
+            label={selectAllOptions[key]}
+            checked={val}
+            onChange={() => {
+              onChange({
+                [question.formInput]: { ...formInputs[question.formInput], [key]: !val },
+              })
+            }}
+            customRef={refs[`${question.formInput}Ref`]}
+          />
+        ))}
+      <br />
+      {formInputs?.[question.formInput]?.other && (
+        <TextInput
+          placeholder="Please Specify"
+          size="small"
+          noOutline
+          value={formInputs?.[toOtherCamelCase(question.formInput)]}
+          onChange={e =>
+            onChange({
+              [toOtherCamelCase(question.formInput)]: e.target.value,
+            })
+          }
+        />
+      )}
+    </>
+  )
+}
+
+export default SelectAll
